[DF-6560] fixed loading of db
Former-commit-id: 77ba57d1a490a162656a9c458eeed5688a2e2563

+++ Sources/Tools/AssetCache/ServerCacheEntry.cpp
@@ -74,7 +74,7 @@ void ServerCacheEntry::Deserialize(KeyedArchive * archieve)
 {
     DVASSERT(nullptr != archieve);
     
-    accessID = archieve->GetUInt32("accessID");
+    accessID = archieve->GetUInt64("accessID");
     
     KeyedArchive *filesArchieve = archieve->GetArchive("files");
     DVASSERT(filesArchieve);
